requirements.txt
fastapi[standard]==0.112.1
pillow==10.4.0
